BE/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
passlib[argon2]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.9
httpx>=0.27.0
pytest>=8.0.0
python-docx>=1.1.2
